+++ build.gradle
@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlinVersion = '1.2.60'
+    ext.kotlinVersion = '1.3.11'
 
     repositories {
         google()
@@ -40,6 +40,7 @@ ext {
     // Network
     okhttpVersion = '3.10.0'
     retrofitVersion = '2.4.0'
+    gsonVersion = '2.8.5'
 
     // Testing
     junitVersion = '4.12'
@@ -87,6 +88,7 @@ ext {
             "com.squareup.retrofit2:retrofit"                   : retrofitVersion,
             "com.squareup.retrofit2:adapter-rxjava2"            : retrofitVersion,
             "com.squareup.retrofit2:converter-simplexml"        : retrofitVersion,
+            "com.google.code.gson:gson"                         : gsonVersion,
 
             // Testing
             "junit:junit"                                       : junitVersion,
