+++ mvicore-plugin-middleware/src/main/java/com/badoo/mvicore/middleware/model/ConnectionData.kt
@@ -1,0 +1,34 @@
+package com.badoo.mvicore.middleware.model
+
+import com.badoo.mvicore.binder.Connection
+
+data class ConnectionData(
+    val from: String,
+    val to: String,
+    val name: String?
+) {
+    constructor(connection: Connection<out Any, out Any>) : this(
+        connection.from.toString(),
+        connection.to.toString(),
+        connection.name
+    )
+}
+
+data class Element(
+    val payload: Any
+)
+
+sealed class Event {
+    data class Bind(
+        val connectionData: ConnectionData
+    ) : Event()
+
+    data class Data(
+        val connection: ConnectionData,
+        val element: Element
+    ) : Event()
+
+    data class Complete(
+        val connection: ConnectionData
+    ) : Event()
+}
