+++ mvicore-plugin/build.gradle
@@ -1,0 +1,60 @@
+buildscript {
+    repositories {
+        jcenter()
+
+    }
+    dependencies {
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$rootProject.ext.kotlinVersion"
+    }
+}
+
+plugins {
+    id "org.jetbrains.intellij" version "0.4.0"
+}
+
+apply plugin: 'java'
+apply plugin: 'maven'
+apply plugin: 'kotlin'
+apply plugin: 'org.jetbrains.dokka'
+apply plugin: 'org.jetbrains.intellij'
+
+group = 'com.github.badoo'
+
+patchPluginXml {
+    sinceBuild '131'
+}
+
+dependencies {
+    def deps = rootProject.ext.deps
+
+    implementation deps('io.reactivex.rxjava2:rxjava')
+    implementation deps('io.reactivex.rxjava2:rxkotlin')
+    implementation deps('com.google.code.gson:gson')
+    implementation project(':mvicore-plugin-middleware')
+
+    testImplementation deps('junit:junit')
+    testImplementation deps('org.jetbrains.kotlin:kotlin-test-junit')
+    testImplementation deps('org.amshove.kluent:kluent')
+}
+
+sourceCompatibility = "1.7"
+targetCompatibility = "1.7"
+
+repositories {
+    jcenter()
+}
+
+task packageSources(type: Jar, dependsOn: 'classes') {
+    classifier = 'sources'
+    from sourceSets.main.allSource
+}
+
+task packageJavadoc(type: Jar, dependsOn: javadoc) {
+    from javadoc.outputDirectory
+    classifier = 'javadoc'
+}
+
+artifacts {
+    archives packageSources
+    archives packageJavadoc
+}
