+++ mvicore-plugin/src/main/resources/META-INF/plugin.xml
@@ -1,0 +1,36 @@
+<idea-plugin version="2">
+    <id>com.badoo.mvicore.plugin</id>
+    <name>MVICore plugin</name>
+    <version>0.0.1</version>
+    <vendor email="andrei.shikov@corp.badoo.com" url="http://www.tech.badoo.com">Badoo</vendor>
+
+    <description><![CDATA[
+    ]]></description>
+
+    <change-notes><![CDATA[
+    ]]>
+    </change-notes>
+
+    <depends>com.intellij.modules.lang</depends>
+
+    <extensions defaultExtensionNs="com.intellij">
+        <!-- Add your extensions here -->
+        <toolWindow
+            id="MVICore"
+            anchor="right"
+            factoryClass="com.badoo.mvicore.plugin.ToolWindowFactory"/>
+    </extensions>
+X
+    <application-components>
+        <!-- Add your application components here -->
+    </application-components>
+
+    <project-components>
+        <!-- Add your project components here -->
+    </project-components>
+
+    <actions>
+        <!-- Add your actions here -->
+    </actions>
+
+</idea-plugin>
