+++ settings.gradle
@@ -1,7 +1,10 @@
 include ':mvicore'
 include ':mvicore-android'
 include ':mvicore-debugdrawer'
+include ':mvicore-plugin-middleware'
+include ':mvicore-plugin'
 include ':mvicore-demo:mvicore-demo-catapi'
 include ':mvicore-demo:mvicore-demo-feature1'
 include ':mvicore-demo:mvicore-demo-feature2'
 include ':mvicore-demo:mvicore-demo-app'
+
